+++ CHANGES.rst
@@ -18,3 +18,13 @@
 - Deprecate
   ``nti.externalization.internalization.register_legacy_search_module``.
   See https://github.com/NextThought/nti.externalization/issues/35
+- Stop ``ext:registerAutoPackageIO`` from registering the legacy
+  class-name based factories by default. If you need class-name based
+  factories, there are two options. The first is to explicitly
+  register ``IClassObjectFactory`` objects in ZCML (we could add a
+  scanning directive to make that more convenient for large numbers of
+  classes), and the second is to set ``register_legacy_search_module``
+  to a true value in the ZCML directive for
+  ``ext:registerAutoPackageIO``. Note that we expect the behaviour of
+  this attribute to change in the near future.
+  See https://github.com/NextThought/nti.externalization/issues/33